+++ CHANGELOG.md
@@ -6,6 +6,8 @@ We follow [Semantic Versions](https://semver.org/).
 
 - You don’t have to use the @inject decorator for nested dependencies;
   it’s only required for top-level calls. Applying @inject to nested dependencies is entirely optional.
+- Breaking changes:
+  - removed `enter` and `shutdown` methods from `AutoScope` class
 
 ## Version 0.32.2